update: setting skeleton page for fetching data
